requirements.txt
# Web Scraping
selenium-wire==5.1.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1

# Data Processing
pandas==2.2.1

# Utils
python-dotenv==1.0.0
tqdm==4.66.2

# Development
pytest==8.0.2
black==24.2.0
flake8==7.0.0
isort==5.13.2 